+++ src/http/header/mod.rs
@@ -1,17 +1,54 @@
 //! A Collection of Header implementations for common HTTP Headers.
 //!
-//! ## Mime
-//!
+//! ## Mime Types
 //! Several header fields use MIME values for their contents. Keeping with the strongly-typed theme,
 //! the [mime] crate is used in such headers as [`ContentType`] and [`Accept`].
 
-use bytes::{Bytes, BytesMut};
 use std::fmt;
 
-pub use self::accept_charset::AcceptCharset;
+use bytes::{Bytes, BytesMut};
+
+// re-export from actix-http
+// - header name / value types
+// - relevant traits for converting to header name / value
+// - all const header names
+// - header map
+// - the few typed headers from actix-http
+// - header parsing utils
 pub use actix_http::http::header::*;
-//pub use self::accept_encoding::AcceptEncoding;
+
+mod accept;
+mod accept_charset;
+mod accept_encoding;
+mod accept_language;
+mod allow;
+mod cache_control;
+mod content_disposition;
+mod content_language;
+mod content_range;
+mod content_type;
+mod date;
+mod encoding;
+mod entity;
+mod etag;
+mod expires;
+mod if_match;
+mod if_modified_since;
+mod if_none_match;
+mod if_range;
+mod if_unmodified_since;
+mod last_modified;
+mod macros;
+mod preference;
+// mod range;
+
+#[cfg(test)]
+pub(crate) use macros::common_header_test;
+pub(crate) use macros::{common_header, common_header_test_module};
+
 pub use self::accept::Accept;
+pub use self::accept_charset::AcceptCharset;
+pub use self::accept_encoding::AcceptEncoding;
 pub use self::accept_language::AcceptLanguage;
 pub use self::allow::Allow;
 pub use self::cache_control::{CacheControl, CacheDirective};
@@ -30,11 +67,10 @@ pub use self::if_none_match::IfNoneMatch;
 pub use self::if_range::IfRange;
 pub use self::if_unmodified_since::IfUnmodifiedSince;
 pub use self::last_modified::LastModified;
+pub use self::preference::Preference;
 //pub use self::range::{Range, ByteRangeSpec};
-pub(crate) use actix_http::http::header::{
-    fmt_comma_delimited, from_comma_delimited, from_one_raw_str,
-};
 
+/// Format writer ([`fmt::Write`]) for a [`BytesMut`].
 #[derive(Debug, Default)]
 struct Writer {
     buf: BytesMut,
@@ -62,30 +98,3 @@ impl fmt::Write for Writer {
         fmt::write(self, args)
     }
 }
-
-mod accept_charset;
-// mod accept_encoding;
-mod accept;
-mod accept_language;
-mod allow;
-mod cache_control;
-mod content_disposition;
-mod content_language;
-mod content_range;
-mod content_type;
-mod date;
-mod encoding;
-mod entity;
-mod etag;
-mod expires;
-mod if_match;
-mod if_modified_since;
-mod if_none_match;
-mod if_range;
-mod if_unmodified_since;
-mod last_modified;
-
-mod macros;
-#[cfg(test)]
-pub(crate) use macros::common_header_test;
-pub(crate) use macros::{common_header, common_header_deref, common_header_test_module};